html/browsers/browsing-the-web/navigating-across-documents/001.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [3]

description: |
  1. Creates an iframe with `src="001-1.html"`
  2. Creates an `onload` handler which will set a new src to the iframe
      (001-1.html => 001-2.html)
  3. 001-1.html has an `unload` handler, that will attempt to set
      `location` to 001-3.html (as a "cross origin" resource, via
      `location.href.replace("http://", "http://www.")`.
      This must NOT be allowed, per https://github.com/whatwg/html/commit/2a11eef8f92b0335328890dc2cb1e281d18553eb.

/-->
<title>Cross-origin navigation from unload handler</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="001-1.html"></iframe>
<script>
var test = async_test();
var iframe = document.querySelector("iframe");

window.onload = test.step_func(function() {
  iframe.src = iframe.src.replace(/\d{3}-\d\.html/, "001-2.html");
});

window.onmessage = test.step_func(function(e) {
  assert_equals(e.data, "001-2");
  test.done();
});
</script>
